chore-LUX(refactor): refactor sendMail function - create main sendEmail html - tests
[Delivers]
