+++ .codeclimate.yml
@@ -2,5 +2,6 @@ version: "2"
 exclude_patterns:
   - "**/node_modules"
   - "**/*.spec.ts"
+   - "**/*.test.js"
   - "**/__mocks__/*"
   - "./src/sequelize/*"
