+++ src/utils/mailer/mainMailerHTML.js
@@ -1,0 +1,13 @@
+export const mainHTML = (message) => `<div style="background:#ECF0F1;width:100%;padding:20px 0;">
+    <div style="max-width:760px;margin:0 auto;background:#ffffff" font-size:1.2em>
+    <div style="background:#266cef;padding:10px;color:#ffffff;text-align:center;font-size:34px">
+    LIG
+    </div>
+    <div style="padding:20px;text-align:left;color:black" font-family: verdana>
+    ${message}
+    </div>
+    </div>
+    <div style="padding:30px 10px;text-align:center;">
+    Copyright &copy; 2020
+    </div>
+    </div>`;
