+++ src/utils/mailer/sendEmail.js
@@ -1,10 +1,11 @@
 import nodemailer from 'nodemailer';
 import { ApolloError } from 'apollo-server-express';
+import { mainHTML } from './mainMailerHTML';
 
 const { NODE_MAILER_USER, NODE_MAILER_PASSWORD, NODE_ENV } = process.env;
 
-
 export const sendEmail = async (email, subject, message) => {
+  const mainHtml = mainHTML(message);
   try {
     const transporter = await nodemailer.createTransport({
       service: 'gmail',
@@ -17,21 +18,9 @@ export const sendEmail = async (email, subject, message) => {
       from: NODE_MAILER_USER,
       to: email,
       subject,
-      html: `<div style="background:#ECF0F1;width:100%;padding:20px 0;">
-      <div style="max-width:760px;margin:0 auto;background:#ffffff" font-size:1.2em>
-      <div style="background:#266cef;padding:10px;color:#ffffff;text-align:center;font-size:34px">
-      LIG
-      </div>
-      <div style="padding:20px;text-align:left;color:black" font-family: verdana>
-      ${message}
-      </div>
-      </div>
-      <div style="padding:30px 10px;text-align:center;">
-      Copyright &copy; 2020
-      </div>
-      </div>`,
+      html: mainHtml,
     };
-    return (NODE_ENV === 'test') ? true : await transporter.sendMail(mailOptions);
+    return NODE_ENV === 'test' ? true : await transporter.sendMail(mailOptions);
   } catch (err) {
     throw new ApolloError(`This has occured during when sending email, ${err}`);
   }
